pyproject.toml
[tool.poetry]
name = "server stats"
version = "0.0.1"
description = "An application for tracking resources consumed by a remote server"
authors = ["ivannikovaleksej"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.2"
psutil = "^5.9.8"
discord-webhook = "^1.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
